tests: move connection map to separate module
To clean up the test_flows module a bit, move the connection map
functionality to a separate module. This is now possible because
beerocks_cli_command can be called directly from the environment module
and no longer needs a TestFlows object.
